refactor prompt gateway
- move
