+++ crates/common/src/lib.rs
@@ -4,10 +4,10 @@ pub mod common_types;
 pub mod configuration;
 pub mod consts;
 pub mod embeddings;
+pub mod errors;
 pub mod http;
 pub mod llm_providers;
 pub mod ratelimit;
 pub mod routing;
 pub mod stats;
 pub mod tokenizer;
-pub mod errors;
