+++ crates/prompt_gateway/src/lib.rs
@@ -3,8 +3,9 @@ use proxy_wasm::traits::*;
 use proxy_wasm::types::*;
 
 mod filter_context;
-mod stream_context;
 mod hallucination;
+mod http_context;
+mod stream_context;
 
 proxy_wasm::main! {{
     proxy_wasm::set_log_level(LogLevel::Trace);
